feat(java): enriquece IndicadorAbrigoResponse com campos do .NET e corrige tipo do score

+++ vesta-api/src/main/java/br/com/fiap/vesta/client/dto/CriticidadeResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record CriticidadeResponse(
     Long idAbrigo,
-    int score,
+    double score,
     String nivel,
     String justificativa,
     List<String> recomendacoes
